+++ BetterReturnScepter/src/ModEntry.cs
@@ -89,11 +89,11 @@ namespace BetterReturnScepter
             // If the return sceptre cooldown is over...
             if (this.rodCooldown.CanWarp)
                 // And if the player is holding a return scepter...
-                if (player.CurrentItem is Wand && player.CurrentItem.Name.Equals("Return Scepter"))
+                if (player.CurrentItem is Wand && player.CurrentItem.ItemId.Equals("ReturnScepter"))
                 {
                     // The cooldown is over, and the player is holding a return sceptre. Now we check to see
                     // whether they press the return to previous point bind, or the Multiple Mini-Obelisk menu bind.
-                    if (e.Pressed.Contains(SButton.MouseRight) || this.config.ReturnToLastPoint.JustPressed())
+                    if ((e.Pressed.Contains(SButton.MouseRight) && this.config.WarpOnRightClick) || this.config.ReturnToLastPoint.JustPressed())
                     {
                         // The player pressed the return to previous point bind.
                         if (!player.bathingClothes.Value && player.IsLocalPlayer && !player.onBridge.Value)
@@ -200,11 +200,9 @@ namespace BetterReturnScepter
 
         private void RegisterWithGmcm()
         {
-            // Get our API reference.
             var configMenuApi =
                 this.Helper.ModRegistry.GetApi<IGenericModConfigMenuApi>("spacechase0.GenericModConfigMenu");
 
-            // If this is null, GMCM wasn't installed.
             if (configMenuApi == null)
             {
                 this.logger.Log("The user doesn't have GMCM installed. This is not an error.");
@@ -212,12 +210,10 @@ namespace BetterReturnScepter
                 return;
             }
 
-            // Register with GMCM
             configMenuApi.Register(this.ModManifest,
                 () => this.config = new ModConfig(),
                 () => this.Helper.WriteConfig(this.config));
 
-            // The toggle for whether or not we want Multiple Mini-Obelisks support.
             configMenuApi.AddBoolOption(
                 this.ModManifest,
                 name: () => "Multiple Mini-Obelisks mod support",
@@ -226,7 +222,6 @@ namespace BetterReturnScepter
                 getValue: () => this.config.EnableMultiObeliskSupport,
                 setValue: value => this.config.EnableMultiObeliskSupport = value);
 
-            // Whether or not to count using MMO's warp menu as using the return scepter.
             configMenuApi.AddBoolOption(
                 this.ModManifest,
                 name: () => "Count warp menu as using scepter",
@@ -235,31 +230,34 @@ namespace BetterReturnScepter
                 getValue: () => this.config.CountWarpMenuAsScepterUsage,
                 setValue: value => this.config.CountWarpMenuAsScepterUsage = value);
 
-            // Add a nice title for prettiness.
+            configMenuApi.AddBoolOption(
+                this.ModManifest,
+                name: () => "Return using right click",
+                tooltip: () =>
+                    "Enable or disable using right click to warp back to your previous spot.",
+                getValue: () => this.config.WarpOnRightClick,
+                setValue: value => this.config.WarpOnRightClick = value);
+
             configMenuApi.AddSectionTitle(
                 this.ModManifest,
                 () => "Keybinds");
 
-            // The bind for opening MMO's warp menu.
             configMenuApi.AddKeybindList(
                 this.ModManifest,
                 name: () => "Open warp menu",
                 getValue: () => this.config.OpenObeliskWarpMenuKbm,
                 setValue: value => this.config.OpenObeliskWarpMenuKbm = value);
 
-            // Add a nice title for prettiness.
             configMenuApi.AddSectionTitle(
                 this.ModManifest,
                 () => "Controller Keybinds");
 
-            // The controller bind for warping to our last point.
             configMenuApi.AddKeybindList(
                 this.ModManifest,
                 name: () => "Warp to last point",
                 getValue: () => this.config.ReturnToLastPoint,
                 setValue: value => this.config.ReturnToLastPoint = value);
 
-            // The controller bind for opening MMO's warp menu.
             configMenuApi.AddKeybindList(
                 this.ModManifest,
                 name: () => "Open warp menu",